This commit refactors the DBDScanner class in the DBDScanner.py file to use a dictionary to build DBD objects instead of directly accessing the JSON response from the pfamscan service.
The new build_DBD_from_dict function is introduced to parse the dictionary and create DBD objects.

The DBD class was updated to take the end instead of the size (the size is calculated) and the accession. The __str__ method was updated accordingly and a getter for the accession was added.